+++ move.c
@@ -14,7 +14,8 @@ static double mod(double x, double y)
 
 void move_entity(game_state_t *state, entity_t *ent)
 {
-	long double delta = state->delta;
+	float delta = state->delta;
+
 	ent->vel.x += delta * ent->acc.x;
 	ent->vel.y += delta * ent->acc.y;
 	ent->pos.x += delta * ent->vel.x;
@@ -48,6 +49,7 @@ void move_box(game_state_t *state, box_t *box)
 						-entity->size.x)
 						+ entity->size.x/2;
 	}
+
 	move_entity(state, entity);
 }
